Rocket Launch Created
When an asteroid is clicked for the first time, a rocket object is
instantiated and launched from Earth. It currently launches from the
same location each time, doesn't rotate, and uses Slerp to catch up to
the asteroid. It also sets the asteroid status to '2' upon collision,
and destroys itself.